src/requirements.txt
Flask
line-bot-sdk
gunicorn
google-genai